Huge Update
Added a new scene "Demo Level" this is the level that we will be showing
to the class on Tuesday. Please do not modify any prefabs or this scene.
If you must test something add it to the "Demo Scene" scene. This can be
found in the scene folding in the assets. I've also added new textures
for the walls and some special events just the demo level.